Refactoring

+++ src/main/java/bot/commands/DeleteCommand.java
@@ -4,7 +4,6 @@ import bot.Commands;
 import bot.commands.interfaces.*;
 import dao.BotDAO;
 import dao.Privacy;
-import dao.WriteToDiskBotDAO;
 
 import static dao.Privacy.CHAT;
 
@@ -13,26 +12,20 @@ public class DeleteCommand extends Command implements Private, Argument, NeedsDA
     private BotDAO dao;
     private String argument;
     private Privacy privacy;
-    private long id;
+    private Long id;
 
-    private String output;
-
-    public DeleteCommand(long chatId) {
+    public DeleteCommand(Long chatId) {
         this.privacy = CHAT;
         this.id = chatId;
     }
 
     @Override
     public void execute() {
+        if (argument == null || id == null || privacy == null || dao == null) throw new IllegalExecutionException();
         boolean successful = dao.remove(argument, id, privacy);
 
-        if (successful) output = "Successfully deleted " + argument;
-        else output = "Couldn't delete " + argument;
-    }
-
-    @Override
-    public String getOutput() {
-        return output;
+        if (successful) setOutput("Successfully deleted " + argument);
+        else            setOutput("Couldn't delete " + argument);
     }
 
     @Override
@@ -59,11 +52,10 @@ public class DeleteCommand extends Command implements Private, Argument, NeedsDA
     @Override
     public String toString() {
         return "DeleteCommand{" +
-                "argument='" + argument + '\'' +
+                "dao=" + dao +
+                ", argument='" + argument + '\'' +
                 ", privacy=" + privacy +
                 ", id=" + id +
-                ", output='" + output + '\'' +
-                '}';
+                "} " + super.toString();
     }
-
 }
